+++ defaults/controllers/_default/root/admin/permissions/new.js
@@ -7,15 +7,9 @@
 
 // action definition
 exports = module.exports = {
-	// functions
+	fieldsExclude: ['type'],
 
-	initForm: function(defaultFn) {
-		return defaultFn().bind(this)
-		.then(function() {
-			// modify form
-			delete this.form.fields.type;
-		});
-	},
+	// functions
 
 	act: function(defaultFn) {
 		return defaultFn().bind(this)